LRSD-8348 [components] Rename DataTable to Table
