+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/components/Table/index.ts
@@ -3,5 +3,5 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default} from './DataTable';
-export type {IColumn, IRow} from './DataTable';
+export {default} from './Table';
+export type {IColumn, IRow} from './Table';
